Add priority to notifications and use the Binary Provider API format
This adds the ability to set a priority to push notifications. In order to have
priority the legacy notifications protocol format had to be deprecated and the
Binary Provider API had to be used. Fortunately, this was mostly a change in the
marshalling implementation.

For more info refer to https://developer.apple.com/library/archive/documentation/NetworkingInternet/Conceptual/RemoteNotificationsPG/BinaryProviderAPI.html#//apple_ref/doc/uid/TP40008194-CH13-SW1